source-kinesis: final change requests
