+++ source-kinesis/capture.go
@@ -21,11 +21,6 @@ import (
 	"golang.org/x/time/rate"
 )
 
-type glueSchemaInfo struct {
-	definition string
-	dataFormat glueTypes.DataFormat
-}
-
 type capture struct {
 	mu     sync.Mutex
 	client *kinesis.Client
@@ -36,7 +31,7 @@ type capture struct {
 	stats map[string]map[string]shardStats
 
 	glueClient      *glue.Client
-	glueSchemaCache map[string]glueSchemaInfo // schemaVersionID -> glueSchemaInfo
+	glueSchemaCache map[string]struct{} // tracks emitted schema version IDs
 }
 
 type shardStats struct {
@@ -434,39 +429,30 @@ func (c *capture) readShard(
 //   - Bytes 18+: Payload
 //
 // Ref: https://github.com/awslabs/aws-glue-schema-registry
-func (c *capture) getGlueSchemaForRecord(ctx context.Context, data []byte) ([]byte, string, glueTypes.DataFormat, bool, error) {
-	if len(data) < 18 {
-		return nil, "", "", false, fmt.Errorf("data too short for Glue header: got %d bytes, need at least 18", len(data))
-	}
-
-	headerVersion := data[0]
-	if headerVersion != 3 {
-		return nil, "", "", false, fmt.Errorf("invalid header version for glue record: got %d expected 3", headerVersion)
-	}
-
+// getGlueSchemaForRecord parses a Glue header from data and returns the JSON Schema if this
+// is a newly-seen schema version. Returns nil if the schema was already cached.
+// The caller should use data[18:] as the payload after this call succeeds.
+func (c *capture) getGlueSchemaForRecord(ctx context.Context, data []byte) (json.RawMessage, error) {
 	compression := data[1]
 	if compression != 0 {
-		return nil, "", "", false, fmt.Errorf("compressed glue schema headers are not supported")
+		return nil, fmt.Errorf("compressed glue schema headers are not supported")
 	}
 
 	if c.glueClient == nil {
-		return nil, "", "", false, fmt.Errorf("got message with glue header, but glue connection previously failed")
+		return nil, fmt.Errorf("got message with glue header, but glue connection previously failed")
 	}
 
-	schemaIDBytes := data[2:18]
-	payload := data[18:]
-
-	schemaUUID, err := uuid.FromBytes(schemaIDBytes)
+	schemaUUID, err := uuid.FromBytes(data[2:18])
 	if err != nil {
-		return nil, "", "", false, fmt.Errorf("parsing schema version UUID: %w", err)
+		return nil, fmt.Errorf("parsing schema version UUID: %w", err)
 	}
 	schemaVersionID := schemaUUID.String()
 
-	// Check cache first
+	// Check cache first - if cached, no schema to emit
 	c.mu.Lock()
-	if cachedInfo, ok := c.glueSchemaCache[schemaVersionID]; ok {
+	if _, ok := c.glueSchemaCache[schemaVersionID]; ok {
 		c.mu.Unlock()
-		return payload, cachedInfo.definition, cachedInfo.dataFormat, true, nil
+		return nil, nil
 	}
 	c.mu.Unlock()
 
@@ -474,196 +460,34 @@ func (c *capture) getGlueSchemaForRecord(ctx context.Context, data []byte) ([]by
 		SchemaVersionId: aws.String(schemaVersionID),
 	})
 	if err != nil {
-		return nil, "", "", false, err
-	}
-
-	schemaDefinition := aws.ToString(out.SchemaDefinition)
-	dataFormat := out.DataFormat
-
-	// Cache the schema
-	c.mu.Lock()
-	c.glueSchemaCache[schemaVersionID] = glueSchemaInfo{definition: schemaDefinition, dataFormat: dataFormat}
-	c.mu.Unlock()
-
-	return payload, schemaDefinition, dataFormat, false, nil
-}
-
-type AvroSchema struct {
-	Type        interface{} `json:"type"`
-	Name        string      `json:"name,omitempty"`
-	LogicalType string      `json:"logicalType,omitempty"`
-	Fields      []struct {
-		Name    string      `json:"name"`
-		Type    interface{} `json:"type"`
-		Default interface{} `json:"default,omitempty"`
-	} `json:"fields,omitempty"`
-	Items   interface{} `json:"items,omitempty"`   // For array types
-	Values  interface{} `json:"values,omitempty"`  // For map types
-	Symbols []string    `json:"symbols,omitempty"` // For enum types
-}
-
-func convertAvroType(avroType interface{}) (map[string]interface{}, bool) {
-	switch t := avroType.(type) {
-
-	case string:
-		return avroTypeToJSONSchema(t, ""), false
-
-	case []interface{}:
-		// Union type - check for nullable pattern ["null", "actualType"]
-		nullable := false
-		var inner map[string]interface{}
-
-		for _, v := range t {
-			if v == "null" {
-				nullable = true
-			} else {
-				inner, _ = convertAvroType(v)
-			}
-		}
-
-		if inner == nil {
-			inner = map[string]interface{}{"type": "string"}
-		}
-		return inner, nullable
-
-	case map[string]interface{}:
-		typeStr, _ := t["type"].(string)
-		logicalType, _ := t["logicalType"].(string)
-
-		switch typeStr {
-		case "array":
-			items, _ := convertAvroType(t["items"])
-			return map[string]interface{}{
-				"type":  "array",
-				"items": items,
-			}, false
-
-		case "map":
-			values, _ := convertAvroType(t["values"])
-			return map[string]interface{}{
-				"type":                 "object",
-				"additionalProperties": values,
-			}, false
-
-		case "enum":
-			symbols, _ := t["symbols"].([]interface{})
-			enumValues := make([]string, 0, len(symbols))
-			for _, s := range symbols {
-				if str, ok := s.(string); ok {
-					enumValues = append(enumValues, str)
-				}
-			}
-			return map[string]interface{}{
-				"type": "string",
-				"enum": enumValues,
-			}, false
-
-		case "fixed":
-			// Fixed bytes - encode as base64 string
-			return map[string]interface{}{
-				"type":            "string",
-				"contentEncoding": "base64",
-			}, false
-
-		case "record":
-			// Nested record - recursively convert
-			fields, _ := t["fields"].([]interface{})
-			properties := map[string]interface{}{}
-			required := []string{}
-
-			for _, f := range fields {
-				field, ok := f.(map[string]interface{})
-				if !ok {
-					continue
-				}
-				fieldName, _ := field["name"].(string)
-				fieldType := field["type"]
-				propSchema, nullable := convertAvroType(fieldType)
-				properties[fieldName] = propSchema
-				if !nullable {
-					_, hasDefault := field["default"]
-					if !hasDefault {
-						required = append(required, fieldName)
-					}
-				}
-			}
-
-			return map[string]interface{}{
-				"type":       "object",
-				"properties": properties,
-				"required":   required,
-			}, false
-
-		default:
-			// Check for logical types on primitive types
-			if logicalType != "" {
-				return avroTypeToJSONSchema(typeStr, logicalType), false
-			}
-			return avroTypeToJSONSchema(typeStr, ""), false
-		}
-	}
-
-	return map[string]interface{}{"type": "string"}, true
-}
-
-// avroTypeToJSONSchema converts an Avro primitive type (with optional logical type) to JSON Schema
-func avroTypeToJSONSchema(avroType string, logicalType string) map[string]interface{} {
-	// Handle logical types first (matching source-kafka behavior)
-	switch logicalType {
-	case "date":
-		return map[string]interface{}{"type": "string", "format": "date"}
-	case "time-millis", "time-micros":
-		return map[string]interface{}{"type": "string", "format": "time"}
-	case "timestamp-millis", "timestamp-micros", "local-timestamp-millis", "local-timestamp-micros":
-		return map[string]interface{}{"type": "string", "format": "date-time"}
-	case "duration":
-		return map[string]interface{}{"type": "string", "format": "duration"}
-	case "uuid":
-		return map[string]interface{}{"type": "string", "format": "uuid"}
-	case "decimal":
-		return map[string]interface{}{"type": "string", "format": "number"}
-	}
-
-	// Handle primitive types
-	switch avroType {
-	case "int", "long":
-		return map[string]interface{}{"type": "integer"}
-	case "float", "double":
-		return map[string]interface{}{"type": "number"}
-	case "boolean":
-		return map[string]interface{}{"type": "boolean"}
-	case "bytes":
-		return map[string]interface{}{"type": "string", "contentEncoding": "base64"}
-	case "null":
-		return map[string]interface{}{"type": "null"}
-	default:
-		return map[string]interface{}{"type": "string"}
-	}
-}
-
-func avroToJSONSchema(schema string) (map[string]interface{}, error) {
-	var avro AvroSchema
-	if err := json.Unmarshal([]byte(schema), &avro); err != nil {
 		return nil, err
 	}
 
-	properties := map[string]interface{}{}
-	required := []string{}
+	schemaDefinition := aws.ToString(out.SchemaDefinition)
 
-	for _, field := range avro.Fields {
-		propSchema, nullable := convertAvroType(field.Type)
-		properties[field.Name] = propSchema
-		// A field is required if it's not nullable AND has no default value
-		if !nullable && field.Default == nil {
-			required = append(required, field.Name)
+	// Cache the schema
+	c.mu.Lock()
+	c.glueSchemaCache[schemaVersionID] = struct{}{}
+	c.mu.Unlock()
+
+	// Convert to JSON Schema based on the data format
+	switch out.DataFormat {
+	case glueTypes.DataFormatAvro:
+		jsonSchemaObj, err := avroToJSONSchema(schemaDefinition)
+		if err != nil {
+			return nil, fmt.Errorf("converting avro to JSON Schema: %w", err)
 		}
-	}
+		return json.Marshal(jsonSchemaObj)
 
-	return map[string]interface{}{
-		"type":       "object",
-		"properties": properties,
-		"required":   required,
-	}, nil
+	case glueTypes.DataFormatJson:
+		if !json.Valid([]byte(schemaDefinition)) {
+			return nil, fmt.Errorf("invalid JSON Schema from Glue")
+		}
+		return json.RawMessage(schemaDefinition), nil
+
+	default:
+		return nil, fmt.Errorf("unsupported Glue schema format: %s", out.DataFormat)
+	}
 }
 
 func (c *capture) processRecords(
@@ -686,56 +510,19 @@ func (c *capture) processRecords(
 
 		// Check for Glue Schema Registry header (version byte = 3)
 		// Header is 18 bytes: 1 version + 1 compression + 16 UUID
-		var payload []byte
-		var glueSchema string
-		var glueDataFormat glueTypes.DataFormat
-		var schemaFromCache bool
-		var err error
+		payload := r.Data
 		if len(r.Data) >= 18 && r.Data[0] == 3 {
-			payload, glueSchema, glueDataFormat, schemaFromCache, err = c.getGlueSchemaForRecord(ctx, r.Data)
+			jsonSchema, err := c.getGlueSchemaForRecord(ctx, r.Data)
 			if err != nil {
 				return fmt.Errorf(
 					"parsing glue header for record with sequenceNumber %q and partitionKey %q: %w",
 					*r.SequenceNumber, *r.PartitionKey, err,
 				)
 			}
-		} else {
-			payload = r.Data
-		}
+			payload = r.Data[18:]
 
-		// If we have a Glue schema that wasn't already emitted, process and emit it
-		if glueSchema != "" && !schemaFromCache {
-			var schemaBytes []byte
-
-			switch glueDataFormat {
-			case glueTypes.DataFormatAvro:
-				// Convert Avro to JSON Schema
-				jsonSchema, err := avroToJSONSchema(glueSchema)
-				if err != nil {
-					return fmt.Errorf("failed to convert avro to JSONSchema: %w", err)
-				}
-				schemaBytes, err = json.Marshal(jsonSchema)
-				if err != nil {
-					return fmt.Errorf("failed to marshal JSON Schema: %w", err)
-				}
-
-			case glueTypes.DataFormatJson:
-				// JSON Schema - use directly, just validate it's valid JSON
-				var jsonSchema map[string]interface{}
-				if err := json.Unmarshal([]byte(glueSchema), &jsonSchema); err != nil {
-					return fmt.Errorf("failed to parse JSON Schema from Glue: %w", err)
-				}
-				schemaBytes = []byte(glueSchema)
-
-			default:
-				log.WithFields(log.Fields{
-					"dataFormat":     glueDataFormat,
-					"sequenceNumber": *r.SequenceNumber,
-				}).Warn("unsupported Glue schema format, skipping schema emission")
-			}
-
-			if schemaBytes != nil {
-				if err := c.stream.SourcedSchema(bindingIndex, schemaBytes); err != nil {
+			if jsonSchema != nil {
+				if err := c.stream.SourcedSchema(bindingIndex, jsonSchema); err != nil {
 					return fmt.Errorf("emitting sourced schema: %w", err)
 				}
 			}
